Introduces a service registry for the incrementalSignal service.
This fixes an issue since, if using modern plugin syntax, plugins run in
their own class loader, meaning that the same class can't be used across
projects. This breaks the incremental signal up per project, using a
WorkerServiceRegistry so that it can be accessed by multiple tasks
within a project.