Add the annotation artifact as an `implementation` dependency instead of `api`.
This prevents misleading results where one could forget to apply the Gradle
plugin. Imagine module B depends on module A. A applies the Gradle plugin, which
adds an `api` dependency for the annotation. In module B you can reference
the annotations due the transitive dependency, but the compiler plugin won't
run, because the Gradle plugin wasn't applied in this module. By using an
`implementation` dependency you cannot reference the annotations in module B
without applying the Gradle plugin.

Fixes #40