Extract CodeGenerator to separate compiler-api module
This introduces a new `compiler-api` artifact that hosts the `CodeGenerator` interface. This is loaded via `ServiceLoader` and allows both Anvil's internal generators as well as 3rd party other generators to wire up their functionality via this new API.

The general design is almost identical to the original `CodeGenerator`, only adding a new `isApplicable` function + `AnvilContext` to allow implementations to indicate if they are applicable. This logic was previously handwritten into `AnvilComponentRegistrar`, and allows moving this logic out externally to implementations.

One generator － `BindingModuleGenerator` － could not fit in this new pattern because it requires a `ClassScanner` instance. This isn't really a blocker though, it can just stay internal and is still added manually. We could explore supporting its needs via the compiler API as well.

All the compiler APIs are guarded by a new 	`@ExperimentalAnvilApi` annotation to allow for breaking changes going forward while this API incubates.