Split anvil compilation test util into underlying `AnvilCompilation`
This allows for custom configuration of the underlying `KotlinCompilation` while keeping the existing `compileAnvil` function for the common case. This also adds a helper function for retrieving anvil-generated files.

Now a fluent usage of this API can look like so:

```kotlin
val compilation = AnvilCompilation()
  .configureAnvil(generateDaggerFactories = true)
  .apply {
    // Not possible before!
    kotlinCompilation.kotlincArguments += "-Xjvm-default=all"
  }

compilation.compile(<sources>) {
  // Do assertions!
}

assertThat(compilation.generatedAnvilFile("com.example", "GeneratedFile").readText()).isEqualTo("...")
```

We could also expand the return type of `compile()` to be an `AnvilCompilationResult` with extra assertions, similar to Google's compile-testing. That's for another day though if there's interest.