Revert publishing changes from build modernization to address publishing failure
There is currently a failure with plugin versioning when publishing manually that appears to be related to ordering/timing issues in the current set-up, around our application of `copy_properties.gradle` in particular. There are also other changes in the new version of the publishing plugin that we should look at more closely and possibly update for, such as --no-parallel no longer being needed.