Add @SingleIn scope and @ForScope qualifier annotations
These 2 optional annotations can be used to avoid having to manually
define scope annotations for each component or qualifier annotations
for each object that can clash with other. Instead all components can
use SingleIn with the same scope marker as the corresponding
MergeComponent/MergeSubcomponent/ContributesSubcomponent annotation uses.
In the same way ForScope can be used with that scope marker to qualify
bindings or provisions in a certain component.