fix sync 'could not get unknown property libs' in gradle-plugin
fixes #744
fixes #757

We had a symlink for `<root>/gradle-plugin/gradle` which pointed to `<root>/gradle`, so that we wouldn't have to maintain two versions of the wrapper files.  That symlink apparently doesn't always work for the `libs.versions.toml` file.

Instead, we can use the normal `from(file('../gradle/libs.version.toml'))` syntax for the catalog, and just symlink the `gradle/wrapper` directory instead.