Move the `:gradle-plugin` project into the main build
Inspired by Molecule [#131](https://github.com/cashapp/molecule/pull/131)

The `:gradle-plugin` project is now part of two builds.

The "build-logic" build creates a version that can be applied to projects in the main build (samples).  This sits in between the normal `build-logic` build and the main build, so that the `:gradle-plugin` project can still consume the convention plugins.

The "main" build creates the version that's actually published to Maven Central and the plugin portal.  The `:gradle-plugin` project no longer needs special treatment when publishing or executing CI.