Add an `all-checks` capstone shard to CI
https://github.com/re-actors/alls-green

One shard requires that all the others are green or skipped.  This single shard can be added as a required check in GitHub's branch protection rules, then never touched again.  When adding a new shard to the workflow, you just add its name to the "needs" list and you're done.

And yes, it supports matrix strategies without needing to maintain a list of all the permutations.