Only apply binary compatibility and dependency-guard to the true root "anvil" build
The `apiDump` and `dependencyGuardBaseline` tasks write to `$projectDir/api` and `$projectDir/dependencies` respectively.  Those directories are shared by the two "anvil" builds (the root build and the included build), so that can cause conflicts.  Fortunately, there's also no point in applying those plugins to the included build since the output would be identical.