Don't cache the `projectDir` or `binaryFile` as part of `GeneratedFileCache`
fixes #880

The `GeneratedFileCache` needs a reference to the current machine's project directory
and the path of its binary file.
These values were being written to that binary file and restored on subsequent reads.
That breaks things when the same cache is used on a different machine due to remote build caching.

Now, those two files are still part of `GeneratedFileCache`,
but only the inner `Tables` class is actually cached.