use the resolved value of `const` arguments in propagated annotation arguments
We have a good deal of logic around parsing out the primitive values of `const`
parameters, but that logic is not recursive.
So, if a `const` references another `const` in its initializer, we wound up
just copying the text of that initializer into the generated annotations,
as though it's just a String.
This problem was isolated to the PSI side of parsing,
since the Descriptor APIs are able to parse out the final, primitive value.
Our type resolution logic always defaults to the PSI models.

Now, when we need to parse out the primitive values from a `PropertyReference`,
we first try to resolve the Descriptor version with a `PropertyDescriptor`.
That isn't possible if the annotation is referencing a `const`
property that was generated in the same round of compilation.
In that event, the parsing will fall back to the PSI models.
Anvil doesn't actually generate a `const`
and then use it in an annotation, so this seems like a reasonable compromise.

fixes #938