delay `@ContributesSubcomponent` generation until the last analysis rounds
The `ContributesSubcomponentHandlerGenerator` is responsible for generating these:

```kotlin
@MergeSubcomponent(scope = Any::class)
public interface SubcomponentInterfaceA : SubcomponentInterface {
  // ...
}
```

It used to generate those interfaces as soon as it scanned the triggers,
but that can cause problems if another `@ContributesSubcomponent` interface is
being generated, and that intends to replace a human-written one.

Now, the handler is a `PrivateCodeGenerator` and it doesn't start parsing/generating until all the "public" code generation is done.  This means it gets the full picture of the possible graph, but it still finishes generating before any IR merging happens.