Always generate provider factories for binding modules.
fixes #948

If full Dagger factory generation is happening in any way
(via Anvil, Dagger with KSP, or Dagger with KAPT),
then behavior doesn't change.
If none of those factory sources are happening in a target compilation,
and that compilation is creating a binding module with a `@Provides`
function (for an `object`), then the `Contribution`
will take it upon itself to generate the factory.

We determine if the Dagger generation is happening
by inspecting the dependencies in the kapt/ksp plugin classpaths.