Gracefully handle module name resolution in KSP (#947)
* Gracefully handle module name resolution in KSP

Due to lack of upstream support in KSP + a bug in the K2 analysis API around exposing module names for mangling, we can't properly support non-public APIs in KSP2. This catches those cases and gracefully offers a error message of advice to make it public instead. Previously this would just hard crash at runtime with a ClassNotFoundException

* Don't generate if mangled without available mangling

* Update message