cache generated file paths relative to the build directory (changed from project directory)
A project's build directory can be reassigned to something other than `<the project>/build`.  This would typically be done via some build logic that's checked in to source control and applied to any machine that builds the code.  This would usually mean that all machines' build directories have the same path relative to the project root.  However, it's possible to build directory to be re-mapped in other ways, that result in a different relative path.  When this happens, the paths for source/generated files must be cached as their paths relative to the build directory, instead of the project.