fix incremental behavior when the compilation target has IR merging but no KAPT stub generation
KGP's "classpath snapshot" incremental logic uses a previous build's `.class` files to track references and invalidate files.  Assume there are two projects: `:lib1` and `:lib2`.  `:lib2` depends on `:lib1`.  If a change is made to an existing type in `:lib1` since the last successful build, KGP will inspect the snapshot of `:lib2`'s class files to see if it references anything in `:lib1`.  If `:lib2` didn't already reference `:lib1`, then the `:lib2:compileKotlin` task will be treated as `UP_TO_DATE` and will not execute again.

Anvil's module and interface merging logic only changes the IR symbols.  This means that all merged references are added to the resultant binary, but not to class files.  Those changes are invisible to the classpath snapshots.  This is why we disable incremental compilation in KAPT stub generation tasks, but it's also a problem in non-KAPT modules that are using other merging annotations.

### The fix

If a compilation merges any interfaces or modules, a file is created with their names.  The `AnvilPlugin` checks for the existence of this file during task configuration, and disables incremental compilation for that task if it's present.  The file won't exist on a fresh build, regardless of whether there will be merging, but that's okay because the build wouldn't be incremental anyway.